[controller] Extend capabilities by making bitmask have 64 bits instead of 16.
Former-commit-id: 10eb434536a950ffbc13400372d3fc0bfabf1cc6
Former-commit-id: df4873e3e6ed9c718e0568bb8e5bb47568261a44

+++ engine/source/controller_delegates.h
@@ -21,7 +21,7 @@ public:
         return controller;
     }
 
-    inline bool has_cap( unsigned int cap )
+    inline bool has_cap( unsigned long long cap )
     {
         return controller->get_capabilities( ) & cap;
     }
